[castai-cluster-controller] Fix volume deployment when not specifying priority class (#666)
* Fix cluster controller deployment when not specifying priorityClass

Currently the definition of the shared-metadata volume in the pod is
guarded by the check for priorityClass. When not specifying this in
the helm values the volume is undefined in the resulting pod
definition but it's still referenced in the volume mounts. Fix this by
moving it out of the if clause.

* bump version

---------

Co-authored-by: Tobias Henkel <tobias.henkel@bmw.de>